chore: ignore examples/pets
This ensures that files in `examples/pets` are only interpreted as plain
source files. In particular, `BUILD.bazel` files under that directory
will not be interpreted by Bazel to create packages and targets.

This solves issues where, if bazel is manually invoked under the
`examples/pets/` WORKSPACE and creates convenience symlinks
(`bazel-out/` etc), the IDE might follow these symlinks and spend a lot
of time looking for pacakges under there.