Stop using `org.pkl-lang:pkl-cli-java`, use `org.pkl-lang:pkl-tools` instead
`org.pkl-lang:pkl-tools` already bundles the same Pkl CLI `Main`
class, so the additional `org.pkl-lang:pkl-cli-java` is unnecessary.

`pkl-cli-java` shell front matter
---------------------------------

The `pkl-cli-java` jar is prefixed with a small amount of shell front
matter to allow it to both be used as a dependency and as an
executable if the correct execution bit is set.

This can possibly cause problems with some tooling, so removing it is
preferred. The front matter:

```
❯ curl -L -o jpkl 'https://repo1.maven.org/maven2/org/pkl-lang/pkl-cli-java/0.26.1/pkl-cli-java-0.26.1.jar'

❯ head -n 2 jpkl
 #!/bin/sh
 exec java  -jar $0 "$@"
```

See: https://pkl-lang.org/main/current/pkl-cli/index.html#java-executable