[bzlmod] Export the `pkl-tools` dep from ruleset to better support `bzlmod`-enabled builds (#68)
This is because of the way that repo visibility works with `bzlmod` enabled
builds. While we can see the `rules_pkl_deps` repo from `rules_pkl` itself,
anything loads `rules_pkl` as a `bazel_dep` cannot. To avoid this problem, we
can re-export the target from `rules_pkl` and make people depend upon that.