fix!: do not read version/name from pkl_project
Previously, we relied on the PklProject name and version fields to be
passed via a provider so that the pkl_project rules could determine at
analysis time the name of output files that will be generated.

Now, pkl_project generates an output directory instead, and that rule is
the one that is responsible to evaluating name and version from the
PklProject file.

This solves a problem we had in the past where the PklProject pkl
content was evalutated twice: once by pkl_project (via `pkl eval`) and
once by pkl_package (via `pkl project package`), sometimes ending up in
different evaluations, if e.g. env vars were different between the two.

This also solves the problem of having to ensure there aren't any
discrepancies between the version evaluated in the PklProject and the
one passed as an attribute to `pkl_package`: the `version` attribute was
removed, and the user can use the `extra_flags` attribute instead, e.g.
`extra_flags = ["--env-var=MY_VERSION=1.2.3"]` (which has the advantage
of working with Pkl properties and any other kind of parametrization
that may arise in Pkl in the future).

BREAKING CHANGES: The provider API changed. A `pkl_package` target now
returns a directory output, and the path to the packaged pkl project
ends up in a subdirectory.