build: reenable integration tests
The integration tests were deadlocking on trying to use the same output
base as the root bazel invocation because of a value set in the home
bazelrc by the Github Actions setup-bazel extension. Now that a
specific, distinct output base is set in a preceding step, integration
tests can be re-enabled.